+++ mxflib/sopsax.cpp
@@ -1,7 +1,7 @@
 /*! \file	sopSAX.cpp
  *	\brief	'sopranino SAX' super-light SAX style XML Parsers
  *
- *	\version $Id: sopsax.cpp,v 1.7 2006/02/11 15:58:38 matt-beard Exp $
+ *	\version $Id: sopsax.cpp,v 1.8 2006/07/02 13:27:51 matt-beard Exp $
  *
  */
 /*
@@ -177,7 +177,7 @@ bool mxflib::sopSAXParseFile(sopSAXHandlerPtr sax, void *UserData, const char *f
 		CurrentAttrib = &AttribBuffer[0];
 		while(attrib < MAXATTRIBS)
 		{
-			int i;
+			size_t i;
 
 			/* Index current attribute's starting point in buffer */
 			Attribs[attrib*2] = CurrentAttrib;
@@ -188,7 +188,7 @@ bool mxflib::sopSAXParseFile(sopSAXHandlerPtr sax, void *UserData, const char *f
 			
 			/* Work out how much buffer is now free */
 			i = strlen(Attribs[attrib*2]) + 1;
-			AttribBufferFree -= i;
+			AttribBufferFree -= static_cast<int>(i);
 			CurrentAttrib += i;
 
 			if(AttribBufferFree < 3)
@@ -265,7 +265,7 @@ bool mxflib::sopSAXParseFile(sopSAXHandlerPtr sax, void *UserData, const char *f
 			
 			/* Work out how much buffer is now free */
 			i = strlen(Attribs[attrib*2+1]) + 1;
-			AttribBufferFree -= i;
+			AttribBufferFree -= static_cast<int>(i);
 			CurrentAttrib += i;
 
 			attrib++;
